refactor: useRecentSearch hook
