+++ components/search/RecentSearch.tsx
@@ -1,11 +1,8 @@
-// 최근 검색어 관련 로직이 담겨있는 컴포넌트
-import { useState } from "react";
-
 import { View, Text, ScrollView, Pressable } from "react-native";
 
 import RecentSearchListButton from "@/components/search/RecentSearchListButton";
 import RecentSearchStatus from "@/components/search/RecentSearchStatus";
-import { setState } from "@vitest/expect";
+import useRecentSearch from "@/hooks/useRecentSearch";
 
 type Props = {
   textInput: string;
@@ -13,22 +10,7 @@ type Props = {
 };
 
 export default function RecentSearch({ textInput, setSearchText }: Props) {
-  // TODO : 최근 검색어 가져오는 기능 구현 필요.
-  const [recentSearchData, setRecentSearchData] = useState<string[]>([
-    "6월 축제",
-    "푸드 페스티벌",
-    "고창 축제",
-    "7월 축제",
-    "8월 축제",
-    "9월 축제",
-  ]);
-  // TODO : 삭제하는 경우 서버에 api 요청을 보내야함.
-  const handleDelete = (keyword: string) => {
-    setRecentSearchData((prev) => prev.filter((item) => item !== keyword));
-  };
-  const handleDeleteAll = () => {
-    setRecentSearchData([]);
-  };
+  const { handleDelete, handleDeleteAll, recentSearchData } = useRecentSearch();
 
   if (textInput.length === 0 && recentSearchData.length > 0) {
     return (
